+++ src/main/java/com/zerobase/hoops/chat/chat/MessageType.java
@@ -1,4 +1,4 @@
-package com.zerobase.hoops.chatYohan.chat;
+package com.zerobase.hoops.chat.chat;
 
 public enum MessageType {
   CHAT,
