--- src/main/java/com/zerobase/hoops/chat/domain/dto/Content.java
@@ -1,9 +1,0 @@
-package com.zerobase.hoops.chat.domain.dto;
-
-import lombok.Getter;
-
-@Getter
-public class Content {
-
-  private String content;
-}
