--- src/main/java/com/zerobase/hoops/chat/domain/dto/CreateRoomDTO.java
@@ -1,16 +1,0 @@
-package com.zerobase.hoops.chat.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-public class CreateRoomDTO {
-
-  private Long gameId;
-
-}
