--- src/main/java/com/zerobase/hoops/chat/domain/repository/ChatRoomRepository.java
@@ -1,10 +1,0 @@
-//package com.zerobase.hoops.chat.domain.repository;
-//
-//import com.zerobase.hoops.entity.ChatRoomEntity;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface ChatRoomRepository extends JpaRepository<ChatRoomEntity, Long> {
-//
-//}
