--- src/main/java/com/zerobase/hoops/chat/domain/repository/MessageRepository.java
@@ -1,10 +1,0 @@
-//package com.zerobase.hoops.chat.domain.repository;
-//
-//import com.zerobase.hoops.entity.MessageEntity;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
-//
-//}
